increasing the size of chunks... to infinity and behond!

+++ install/brender.sql
@@ -58,7 +58,7 @@ CREATE TABLE IF NOT EXISTS `jobs` (
   `end` int(11) NOT NULL DEFAULT '100',
   `project` varchar(32) NOT NULL,
   `current` int(11) NOT NULL DEFAULT '0',
-  `chunks` tinyint(4) NOT NULL DEFAULT '0',
+  `chunks` int(11) NOT NULL DEFAULT '0',
   `filetype` varchar(16) NOT NULL DEFAULT 'PNG',
   `rem` varchar(255) NOT NULL,
   `config` varchar(64) NOT NULL,
